fix(earnings): Enforce financial accuracy and refactor payout workflow
The earnings module was refactored to address critical financial inconsistencies and architectural debt.

### Key Fixes:

1.  **Financial Accuracy (BigDecimal):**
    * Updated all Driver Earnings DTOs (DriverEarningsDTO, DriverPayoutDTO, EarningsSummaryDTO) to use java.math.BigDecimal for all currency fields, eliminating floating-point errors from double conversions.
    * Refactored EarningsServiceImpl to perform all money summation and calculations using BigDecimal::add and pass BigDecimals directly to DTOs.
2.  **Payout Workflow:**
    * Added status (as PayoutStatus Enum), 	ransactionRef, and processedBy fields to the DriverPayout domain entity to support a proper state machine (PENDING, PAID).
    * Modified equestManualPayout to immediately set the status to PENDING, deferring the actual deposit into the driver's wallet until processPayout is called by an administrator, correcting the previous instant deposit bug.
    * Updated processPayout to handle status checks and record transaction details.
3.  **Architectural Cleanup:**
    * Removed the redundant DriverPayoutController to consolidate all driver-facing earnings API endpoints under the main EarningsController.

+++ backend/src/main/java/com/deliverXY/backend/NewCode/earnings/dto/DriverEarningsDTO.java
@@ -3,12 +3,14 @@ package com.deliverXY.backend.NewCode.earnings.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class DriverEarningsDTO {
     private Long deliveryId;
-    private Double earnings;
-    private Double tip;
+    private BigDecimal earnings;
+    private BigDecimal tip;
     private String createdAt;
 }
