+++ backend/src/main/java/com/deliverXY/backend/NewCode/earnings/dto/EarningsSummaryDTO.java
@@ -3,11 +3,13 @@ package com.deliverXY.backend.NewCode.earnings.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class EarningsSummaryDTO {
-    private double totalEarned;
-    private double totalTips;
-    private double totalDeliveries;
+    private BigDecimal totalEarned;
+    private BigDecimal totalTips;
+    private Long totalDeliveries;
     private double totalDistanceKm;
 }
